RAG/pyproject.toml
[project]
name = "langgraph-tutorial-mod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ansi2html>=1.9.2",
    "colorama>=0.4.6",
    "gradio>=5.34.1",
    "graphviz>=0.20.3",
    "h5py>=3.14.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langchain-groq>=0.3.2",
    "langchain-pinecone>=0.2.8",
    "langgraph>=0.4.8",
    "langsmith>=0.3.45",
    "matplotlib>=3.10.3",
    "nltk>=3.9.1",
    "omegaconf>=2.3.0",
    "openai>=1.84.0",
    "pinecone>=7.0.2",
    "pinecone-text>=0.5.4",
    "pygraphviz>=1.14",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=4.1.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.52.4",
]

[tool.uv.workspace]
members = [
    "myRAg/PDF-Extract-Kit/scripts",
]
